Converted profile to server component
