+++ client/.env.example
@@ -1,1 +1,2 @@
 NEXT_PUBLIC_ACM_API_URL="http://localhost:4000/api/v1"
+NEXT_PUBLIC_BASE_URL="http://localhost:3000"
